refactored to keep builder methods in the expected files
