+++ x509-ocsp/src/responder_id.rs
@@ -1,0 +1,48 @@
+//! X.509 OCSP ResponderID
+
+use der::{asn1::OctetString, Choice};
+use x509_cert::name::Name;
+
+/// ResponderID structure as defined in [RFC 6960 Section 4.2.1].
+///
+/// ```text
+/// ResponderID ::= CHOICE {
+///    byName              [1] Name,
+///    byKey               [2] KeyHash }
+/// ```
+///
+/// [RFC 6960 Section 4.2.1]: https://datatracker.ietf.org/doc/html/rfc6960#section-4.2.1
+#[derive(Clone, Debug, Eq, PartialEq, Choice)]
+#[allow(missing_docs)]
+pub enum ResponderId {
+    #[asn1(context_specific = "1", tag_mode = "EXPLICIT", constructed = "true")]
+    ByName(Name),
+
+    #[asn1(context_specific = "2", tag_mode = "EXPLICIT", constructed = "true")]
+    ByKey(KeyHash),
+}
+
+impl From<Name> for ResponderId {
+    fn from(other: Name) -> Self {
+        Self::ByName(other)
+    }
+}
+
+impl From<KeyHash> for ResponderId {
+    fn from(other: KeyHash) -> Self {
+        Self::ByKey(other)
+    }
+}
+
+/// KeyHash structure as defined in [RFC 6960 Section 4.2.1].
+///
+/// ```text
+/// KeyHash ::= OCTET STRING -- SHA-1 hash of responder's public key
+///                          -- (i.e., the SHA-1 hash of the value of the
+///                          -- BIT STRING subjectPublicKey [excluding
+///                          -- the tag, length, and number of unused
+///                          -- bits] in the responder's certificate)
+/// ```
+///
+/// [RFC 6960 Section 4.2.1]: https://datatracker.ietf.org/doc/html/rfc6960#section-4.2.1
+pub type KeyHash = OctetString;
